requirements.txt
streamlit==1.37.1
chromadb==0.5.3
sentence-transformers==3.0.1
pypdf==4.2.0
python-dotenv==1.0.1
requests==2.32.3